+++ types/bplist-parser.d.ts
@@ -1,3 +1,0 @@
-declare module "bplist-parser" {
-	export const parseFile: (filename: string) => Promise<any>
-}